refactor: split legacy usage insight queries
